requirements.txt
tensorflow>=2.12.0
keras>=2.12.0
numpy>=1.23.0
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0
scikit-learn>=1.2.0
streamlit>=1.28.0
plotly>=5.15.0
Pillow>=9.4.0
